adds fetchallwalls
