+++ schema/type.js
@@ -7,6 +7,17 @@ const {
   GraphQLString
 } = require("graphql");
 
+exports.ArtistType = new GraphQLObjectType({
+  name: "Artist",
+  type: "Query",
+  description: "One artist",
+  fields: {
+    artist_id: { type: GraphQLID },
+    artist_username: { type: GraphQLString },
+    bio: { type: GraphQLString }
+  }
+});
+
 exports.WallType = new GraphQLObjectType({
   name: "Wall",
   type: "Query",
